requirement.txt
django==3.0.6
paho-mqtt
psycopg2-binary>=2.8